requirements.txt
aiohttp==3.9.5
aiohappyeyeballs==2.4.3
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
dateparser==1.2.0
deep-translator==1.11.4
Deprecated==1.2.14
distro==1.9.0
durationpy==0.9
exceptiongroup==1.2.2
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
langchain==0.3.7
langchain-chroma==0.1.4
langchain-community==0.3.3
langchain-core==0.3.15
langchain-openai==0.2.3
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.137
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.12.1
numpy==1.26.4
oauthlib==3.2.2
ocrmac==0.1.7
onnxruntime==1.20.0
openai==1.52.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pinecone==5.3.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pip==24.3.1
posthog==3.7.0
propcache==0.2.0
protobuf==4.25.5
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.0
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreML==10.3.1
pyobjc-framework-Quartz==10.3.1
pyobjc-framework-Vision==10.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
setuptools==65.5.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
streamlit==1.39.0
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.66.5
translator==0.0.9
typer==0.12.5
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvloop==0.21.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
yarl==1.16.0
zipp==3.20.2
